requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.10.4
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
discord.py==2.3.2
frozenlist==1.5.0
google-ai-generativelanguage==0.4.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
rsa==4.9
schedule==1.2.0
six==1.16.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.17.1
